Move time management to separate module
