+++ src/stonefish/mod.rs
@@ -3,8 +3,7 @@ mod evaluation;
 mod heuristic;
 mod node;
 mod types;
-
-use std::time::Duration;
+mod time_management;
 
 use pleco::Board;
 
@@ -17,7 +16,7 @@ use crate::{
     },
 };
 
-use self::types::RepetitionTable;
+use self::{types::RepetitionTable, time_management::get_max_time};
 
 #[derive(Debug, Clone)]
 pub struct Stonefish {
@@ -158,40 +157,9 @@ impl UciEngine for Stonefish {
     fn go(&mut self, go_config: UciGoConfig, stop_flag: AbortFlag) {
         let mut root = Node::new(self.board.clone());
 
-        // Determine search depth if one is given
-        let max_depth = if let Some(max_depth) = go_config.max_depth {
-            Some(max_depth)
-        } else {
-            go_config.search_mate
-        };
-
-        // Determine player time and increment
-        let (time, increment) = match root.board.turn() {
-            pleco::Player::White => (go_config.white_time_ms, go_config.white_increment_ms),
-            pleco::Player::Black => (go_config.black_time_ms, go_config.black_increment_ms),
-        };
-
-        // Determine maximum time
-        let max_time = if let Some(move_time_ms) = go_config.move_time_ms {
-            Some(Duration::from_millis(move_time_ms.try_into().unwrap()))
-        } else if go_config.infinite {
-            // Search infinitely
-            None
-        } else if let Some(time_ms) = time {
-            // Take 5 seconds reserve time for each move
-            let base_time_ms: u64 = 5000.min(time_ms.try_into().unwrap());
-            // Additionally use the increment time
-            let increment_time_ms: u64 = increment.try_into().unwrap();
-            let mut total_time_ms = base_time_ms + increment_time_ms;
-            // Consider a delay of 500 ms and cap at 10 seconds
-            total_time_ms = total_time_ms.saturating_sub(500).min(10000);
-            Some(Duration::from_millis(total_time_ms))
-        } else if max_depth.is_some() {
-            None
-        } else {
-            // Search for 10 seconds
-            Some(Duration::from_millis(10000))
-        };
+        // Determine search depth and time
+        let max_depth = go_config.max_depth.or(go_config.search_mate);
+        let max_time = get_max_time(go_config, root.board.turn());
 
         // Search for the best move
         root.iterative_deepening(
